refactor(server-cloudflare): rename env.ts to hono.ts, add validated env and drizzle config
Rename src/env.ts → src/hono.ts (it's a Hono factory, not env config).
Add src/env.ts with arktype-validated .dev.vars loading, shared by
drizzle.config.ts and better-auth.config.ts. Replace auth:migrate with
drizzle-kit db:push/generate/migrate/studio scripts.
