+++ packages/server-cloudflare/src/proxy/chat.ts
@@ -3,7 +3,7 @@ import {
 	PROVIDER_ENV_VARS,
 	type SupportedProvider,
 } from '@epicenter/sync-core';
-import { factory } from '../env';
+import { factory } from '../hono';
 
 /** Provider API chat completion endpoints. */
 const PROVIDER_CHAT_URL: Record<SupportedProvider, string> = {
